@@ -16,33 +16,35 @@
   "author": "Taylor Denouden",
   "license": "Unlicense",
   "devDependencies": {
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^3.1.0",
-    "eslint-plugin-react": "^3.15.0",
-    "jasmine": "^2.4.1",
-    "jasmine-core": "^2.4.1",
-    "jsdoc-to-markdown": "^1.3.3",
-    "karma": "^0.13.19",
-    "karma-chrome-launcher": "^0.2.2",
-    "karma-jasmine": "^0.3.6",
-    "karma-phantomjs-launcher": "^0.2.3",
-    "karma-requirejs": "^0.2.3",
-    "karma-webpack": "^1.7.0",
-    "node-sass": "^3.4.2",
-    "phantomjs": "^1.9.19",
-    "phantomjs-polyfill": "0.0.1",
-    "requirejs": "^2.1.22",
-    "sass-loader": "^3.1.2",
-    "style-loader": "^0.13.0",
-    "webpack": "^1.12.11"
+    "babel-core": "^6.14.0",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "css-loader": "^0.25.0",
+    "eslint": "^3.5.0",
+    "eslint-config-airbnb": "^11.1.0",
+    "eslint-plugin-import": "^1.15.0",
+    "eslint-plugin-jsx-a11y": "^2.2.2",
+    "eslint-plugin-react": "^6.2.1",
+    "jasmine": "^2.5.1",
+    "jasmine-core": "^2.5.1",
+    "jsdoc-to-markdown": "^1.3.7",
+    "karma": "^1.3.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-jasmine": "^1.0.2",
+    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-requirejs": "^1.0.0",
+    "karma-webpack": "^1.8.0",
+    "node-sass": "^3.10.0",
+    "phantomjs": "^2.1.7",
+    "phantomjs-polyfill": "0.0.2",
+    "requirejs": "^2.3.1",
+    "sass-loader": "^4.0.2",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.2"
   },
   "dependencies": {
-    "simple-statistics": "^1.0.1",
-    "d3": "^3.5.12"
+    "d3": "^4.2.3",
+    "simple-statistics": "^2.1.0"
   }
 }
